CPDSS3: CargoNomination listing changes (#41)
* CPDSS3: fixed merge issue in development branch

* spotlessApply changes

* CPDSS3: cargo nomination loading port mapping changes

* CPDSS3: updated review comments

* CPDSS3: CargoNominations listing

+++ api/cp-dss-api/gateway/src/main/java/com/cpdss/gateway/domain/CargoNomination.java
@@ -17,7 +17,7 @@ public class CargoNomination {
   private Long loadableStudyId;
 
   @NotNull(message = CommonErrorCodes.E_HTTP_BAD_REQUEST)
-  private Integer priority;
+  private Long priority;
 
   private String color;
